adds rules node; for integer type assertions from terraform, checks the value is > 0 rather than isNil

+++ ol_schema/app/parameters/parameter_test.go
@@ -46,7 +46,7 @@ func TestInflateParameter(t *testing.T) {
 				"include_in_saml_assertion":  true,
 			},
 			ExpectedOutput: apps.AppParameters{
-				ID:                        oltypes.Int32(123),
+				ID:                        oltypes.Int32(int32(123)),
 				Label:                     oltypes.String("test"),
 				UserAttributeMappings:     oltypes.String("test"),
 				UserAttributeMacros:       oltypes.String("test"),
